rearrangement of code
